eye-tracking-software/requirements.txt
mediapipe>=0.10.9
opencv-python>=4.9.0
scikit-learn>=1.4.0
numpy>=1.26.0
pyobjc-core>=10.1
pyobjc-framework-Cocoa>=10.1
pyobjc-framework-Quartz>=10.1
pynput>=1.7.6
py2app>=0.28
